backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
google-generativeai==0.8.3
requests==2.32.3
python-dotenv==1.0.1
pydantic==2.10.3
firebase-admin==6.6.0
httpx==0.28.1
google-cloud-texttospeech==2.14.1
